requirements.txt
chardet
impacket
pycryptodome